move organization tests to actions package
