+++ application/actions/gql_test.go
@@ -1,7 +1,6 @@
 package actions
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
@@ -16,171 +15,6 @@ type gqlErrorResponse struct {
 		Path    []string `json:"path"`
 	} `json:"errors"`
 	Data interface{} `json:"data"`
-}
-
-func (as *ActionSuite) Test_CreateOrganization() {
-	t := as.T()
-
-	// Array indexes for convenience in references
-	const (
-		SalesAdmin    = 0
-		OrgMember     = 1
-		OrgAdmin      = 2
-		OtherOrgAdmin = 3
-		Org1          = 0
-	)
-
-	testFixtures := Fixtures_CreateOrganization(as, t)
-	userFixtures := testFixtures.Users
-	orgFixtures := testFixtures.Orgs
-
-	type testCase struct {
-		Name            string
-		Token           string
-		Payload         string
-		ExpectError     bool
-		ExpectSubString string
-	}
-
-	createOrgPayload := `{"query": "mutation { createOrganization(input: { name: \"new org\", url: \"http://test.com\", authType: \"saml2\", authConfig: \"{}\", }){id} }"}`
-	updateOrgPayload := fmt.Sprintf(`{"query": "mutation { updateOrganization(input: { id: \"%s\" name: \"updated org\", url: \"http://test.com\", authType: \"saml2\", authConfig: \"{}\", }){id} }"}`, orgFixtures[Org1].UUID.String())
-	createOrgDomainPayload := fmt.Sprintf(`{"query": "mutation { createOrganizationDomain(input: { organizationID: \"%s\", domain: \"newdomain.com\"}){domain} }"}`, orgFixtures[Org1].UUID.String())
-	removeOrgDomainPayload := fmt.Sprintf(`{"query": "mutation { removeOrganizationDomain(input: { organizationID: \"%s\", domain: \"newdomain.com\"}){domain} }"}`, orgFixtures[Org1].UUID.String())
-
-	testCases := []testCase{
-		{
-			Name:        "org member cannot create org",
-			Token:       userFixtures[OrgMember].Nickname,
-			Payload:     createOrgPayload,
-			ExpectError: true,
-		},
-		{
-			Name:        "org admin cannot create org",
-			Token:       userFixtures[OrgAdmin].Nickname,
-			Payload:     createOrgPayload,
-			ExpectError: true,
-		},
-		{
-			Name:            "sales admin can create org",
-			Token:           userFixtures[SalesAdmin].Nickname,
-			Payload:         createOrgPayload,
-			ExpectError:     false,
-			ExpectSubString: "createOrganization",
-		},
-		{
-			Name:        "org member cannot update org",
-			Token:       userFixtures[OrgMember].Nickname,
-			Payload:     updateOrgPayload,
-			ExpectError: true,
-		},
-		{
-			Name:            "org admin can update org",
-			Token:           userFixtures[OrgAdmin].Nickname,
-			Payload:         updateOrgPayload,
-			ExpectError:     false,
-			ExpectSubString: "updateOrganization",
-		},
-		{
-			Name:        "other org admin cannot update org1",
-			Token:       userFixtures[OtherOrgAdmin].Nickname,
-			Payload:     updateOrgPayload,
-			ExpectError: true,
-		},
-		{
-			Name:            "sales admin can update org",
-			Token:           userFixtures[SalesAdmin].Nickname,
-			Payload:         updateOrgPayload,
-			ExpectError:     false,
-			ExpectSubString: "updateOrganization",
-		},
-		{
-			Name:        "org member cannot create org domain",
-			Token:       userFixtures[OrgMember].Nickname,
-			Payload:     createOrgDomainPayload,
-			ExpectError: true,
-		},
-		{
-			Name:            "org admin can create org domain",
-			Token:           userFixtures[OrgAdmin].Nickname,
-			Payload:         createOrgDomainPayload,
-			ExpectError:     false,
-			ExpectSubString: "createOrganizationDomain",
-		},
-		{
-			Name:        "org admin cannot create duplicate org domain",
-			Token:       userFixtures[OrgAdmin].Nickname,
-			Payload:     createOrgDomainPayload,
-			ExpectError: true,
-		},
-		{
-			Name:        "org member cannot remove org domain",
-			Token:       userFixtures[OrgMember].Nickname,
-			Payload:     removeOrgDomainPayload,
-			ExpectError: true,
-		},
-		{
-			Name:            "org admin can remove org domain",
-			Token:           userFixtures[OrgAdmin].Nickname,
-			Payload:         removeOrgDomainPayload,
-			ExpectError:     false,
-			ExpectSubString: "removeOrganizationDomain",
-		},
-		{
-			Name:        "other org admin cannot create org1 domain",
-			Token:       userFixtures[OtherOrgAdmin].Nickname,
-			Payload:     createOrgDomainPayload,
-			ExpectError: true,
-		},
-		{
-			Name:            "sales admin can create org domain",
-			Token:           userFixtures[SalesAdmin].Nickname,
-			Payload:         createOrgDomainPayload,
-			ExpectError:     false,
-			ExpectSubString: "createOrganizationDomain",
-		},
-		{
-			Name:            "sales admin can remove org domain",
-			Token:           userFixtures[SalesAdmin].Nickname,
-			Payload:         removeOrgDomainPayload,
-			ExpectError:     false,
-			ExpectSubString: "removeOrganizationDomain",
-		},
-	}
-
-	for _, tc := range testCases {
-
-		payload := bytes.NewReader([]byte(tc.Payload))
-		req := httptest.NewRequest("POST", "/gql", payload)
-		resp := httptest.NewRecorder()
-		req.Header.Set("Authorization", fmt.Sprintf("Bearer %s", tc.Token))
-		req.Header.Set("content-type", "application/json")
-
-		as.App.ServeHTTP(resp, req)
-
-		body, err := ioutil.ReadAll(resp.Body)
-		if err != nil {
-			t.Error(err)
-		}
-
-		if tc.ExpectError {
-			var errResp gqlErrorResponse
-			err = json.Unmarshal(body, &errResp)
-			if err != nil {
-				t.Error(err)
-			}
-
-			if len(errResp.Errors) == 0 {
-				t.Errorf("did not get expected errors in test %s, response: +%v", tc.Name, errResp)
-			}
-
-			continue
-		}
-
-		if !strings.Contains(string(body), tc.ExpectSubString) {
-			t.Errorf("substring \"%s\" not found in response: %s. test case: %s", tc.ExpectSubString, string(body), tc.Name)
-		}
-
-	}
 }
 
 func (as *ActionSuite) testGqlQuery(gqlQuery, accessToken string, response interface{}) error {
